build-for-compare: Respect TMPDIR
Currently the script hardcodes `/tmp` as temporary directory
for building.

This is not always the correct assumption, for example this might point
to a limited in-mmemory file system in some environments, whereas
the build requires a lot of space.

Change to use `tempfile.gettempdir`. Adapt the safe path checking
to check whether inside this path.