Add github-merge.py
In the IRC meeting of today (2019-08-15) it was discussed that the
github-merge tool is used by various other projects as well. For this,
and other reasons (e.g. making the tool independent of any specific
bitcoin release) It makes sense to move it from the bitcoin core
repository (`contrib/devtools/github-merge.py`) to a more general place.