github-merge: Prevent git logs from being launched in pager
The script is supposed to print these logs inline. The default,
howerver, for `git log` is to pipe them to a pager (such as `less`).
This can result in horrible usability (imo) when there are a lot of commits
or a small terminal is used.

Add an explicit flag `--no-pager` to prevent this from happening. If
there is long output, use the scrollback feature of your terminal.