gitian-verify: Switch to gpg module
Use the 'official' Python gpg module (which interfaces to gnupg through
the gpgme library instead of a subprocess) to verify detached gitian
signatures.

Besides being more robust than the ad-hoc textual parser, this adds
extra functionality: a separate status when a key has expired.

Often this library is already installed, at least on Debian
(and part of the trusted set), so I don't think the extra dependency is
harmful.