list-pulls.py: Minimal support for multi-repository
This adds minimal support for handling the `bitcoin/bitcoin#1234` syntax
in merge messages, as well as scaffolding for handling a project spanning
multiple repositories.

We do not currently have a meta-data backup repository covering
`bitcoin-core/gui` so those PRs are considered unknown but belonging to
the GUI category, for now.